# internal/server/openapi.yaml
openapi: "3.0.0"
info:
  version: 3.0.0
  title: Peregrine API documentation
servers:
  - url: http://edge.api.peregrine.ga:8080/
  - url: http://api.peregrine.ga:8080/
  - url: http://localhost:8080/
paths:
...
  /teams/{teamKey}:
    parameters:
      - $ref: "#/components/parameters/teamKey"
    get:
      summary: Get general info for a specific team
      operationId: getTeam
      tags:
        - teams
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/team"
        "404":
          $ref: "#/components/responses/notFoundError"
        "500":
          $ref: "#/components/responses/internalServerError"
...
  /realms:
    get:
      summary: Get all realms
      operationId: getRealms
      security:
        - BearerAuth: []
      tags:
        - realms
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/realm"
        "401":
          $ref: "#/components/responses/unauthorizedError"
        "500":
          $ref: "#/components/responses/internalServerError"
    post:
      summary: Create a new realm
      description: Only global admins can create realms.
      operationId: createRealm
      tags:
        - realms
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/realm"
      responses:
        "201":
          description: Successfully created realm
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/realm"
        "401":
          $ref: "#/components/responses/unauthorizedError"
        "403":
          $ref: "#/components/responses/forbiddenError"
        "409":
          $ref: "#/components/responses/conflictError"
        "422":
          $ref: "#/components/responses/unprocessableEntityError"
        "500":
          $ref: "#/components/responses/internalServerError"
  /realms/{id}:
    get:
      summary: Get a realm by ID
      operationId: getRealm
      security:
        - BearerAuth: []
      tags:
        - realms
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/realm"
        "401":
          $ref: "#/components/responses/unauthorizedError"
        "403":
          $ref: "#/components/responses/forbiddenError"
        "409":
          $ref: "#/components/responses/conflictError"
        "422":
          $ref: "#/components/responses/unprocessableEntityError"
        "500":
          $ref: "#/components/responses/internalServerError"
    post:
      summary: Update a realm
      description: Global admins can update any realm. Realm admins can update their realm.
      operationId: updateRealm
      tags:
        - realms
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/realm"
      responses:
        "204":
          description: Successfully updated realm
        "401":
          $ref: "#/components/responses/unauthorizedError"
        "403":
          $ref: "#/components/responses/forbiddenError"
        "404":
          $ref: "#/components/responses/notFoundError"
        "422":
          $ref: "#/components/responses/unprocessableEntityError"
        "500":
          $ref: "#/components/responses/internalServerError"
    delete:
      summary: Delete a realm
      operationId: deleteRealm
      security:
        - BearerAuth: []
      tags:
        - realms
      responses:
        "204":
          description: Successfully deleted realm
        "401":
          $ref: "#/components/responses/unauthorizedError"
        "403":
          $ref: "#/components/responses/forbiddenError"
        "409":
          $ref: "#/components/responses/conflictError"
        "422":
          $ref: "#/components/responses/unprocessableEntityError"
        "500":
          $ref: "#/components/responses/internalServerError"
components:
  parameters:
    teamKey:
      in: path
      name: teamKey
      schema:
        $ref: "#/components/schemas/teamKey"
      required: true
      description: Team Key
...
  responses:
    internalServerError:
      description: Failed due to an internal server error
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/internalServerError"
    unauthorizedError:
      description: Invalid access token, please obtain a new access token
      content:
        text/plain:
          schema:
            type: string
            example: Unauthorized
    forbiddenError:
      description: Your access token is valid, but you lack the roles to perform this operation
      content:
        text/plain:
          schema:
            type: string
            example: Forbidden
    notFoundError:
      description: Unable to find that resource
      content:
        text/plain:
          schema:
            type: string
            example: Not Found
    unprocessableEntityError:
      description: Request body syntax was invalid
      content:
        text/plain:
          schema:
            type: string
            example: Unprocessable Entity
    conflictError:
      description: A resource with a similar unique key exists (e.g. username or id)
      content:
        text/plain:
          type: string
          example: Conflict
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    teamKey:
      type: string
      example: frc2733
...
    realm:
      required:
        - name
        - shareReports
        - id
      properties:
        name:
          type: string
          example: Pigmice
        shareReports:
          type: boolean
          example: true
        id:
          $ref: "#/components/schemas/id"
...
    team:
      required:
        - key
      properties:
        key:
          type: string
          example: frc2733
        nickname:
          type: string
          example: RAGE Robotics ⚙️
    id:
      description: Auto-increment 64-bit integer that identifies a resource
      type: integer
      format: int64
      example: 1
      readOnly: true
...
    internalServerError:
      type: string
      example: Internal Server Error
...